send generated pass via email
because I don't want to have to write to temporary files, this requires
a custom patched version of node-passbook which contains a fix for
non-file writable streams.

As it stands right now, you can execute the script and pipe it a
complete email via stdin. It will then produce the pass and send it to
the original recipient